Se corrige un bug en donde se borraba el archivo previo durante la edición, cuando  el cv no había sido editado, perdiendo el único archivo que se tenía.

+++ src/main/java/org/nahual/utils/CvUploader.java
@@ -68,7 +68,7 @@ public class CvUploader implements Upload.Receiver, Upload.SucceededListener {
             } else {
                 // Si existía un cv anterior, se reemplaza por el nuevo
                 cv = new File(cvUtils.getCvPath(egresado));
-                if (cv.exists())
+                if (cv.exists() && filename != null)
                     cv.delete();
             }
